requirements.txt
openai>=1.29.0
pyautogui>=0.9.54
pytesseract>=0.3.10
python-dotenv>=1.0.0
pillow>=10.0.0
